Extract RecordFilter Mixin into compat
