+++ lib/fluent/compat/record_filter_mixin.rb
@@ -1,0 +1,34 @@
+#
+# Fluentd
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+#
+
+module Fluent
+  module Compat
+    module RecordFilterMixin
+      def filter_record(tag, time, record)
+      end
+
+      def format_stream(tag, es)
+        out = ''
+        es.each {|time,record|
+          tag_temp = tag.dup
+          filter_record(tag_temp, time, record)
+          out << format(tag_temp, time, record)
+        }
+        out
+      end
+    end
+  end
+end
